finished repository 40% service

+++ springboot-backend/h3clothing/src/main/java/net/aptech/h3clothing/security/UserDetailServiceImpl.java
@@ -24,7 +24,7 @@ public class UserDetailServiceImpl implements UserDetailsService {
     }
 
     @Transactional
-    public UserDetails loadUserById(Long id) {
+    public UserDetails loadUserById(Integer id) {
         User user = repository.findById(id).orElseThrow(
                 () -> new UsernameNotFoundException("User not found with id : " + id)
         );
